genai-requirements-automation/genai-requirements-automation/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Pydantic
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Document processing
python-docx>=1.1.0
PyPDF2>=3.0.1
pdfplumber>=0.10.3
reportlab>=4.0.7

# Vector store and embeddings
chromadb>=0.4.18
sentence-transformers>=2.2.2

# LLM
openai>=1.3.0

# Azure DevOps
azure-devops>=7.1.0

# Utilities
aiofiles>=23.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

